Styled Navbar / Sidebar
Message:
I have added a styled navbar and sidebar with help from a youtube tutorial "Coding2go".
I have added the code with minor changes to ensure it acts accordingly.
As it now acts as it should. I will now apply my own styling and make it my own for Palms and Peaks.
The navbar is responsive and the sidebar is hidden until the user clicks the hamburger icon.
It then appears from the right side of the screen with an added transition.

	modified:   assets/css/style.css
	modified:   index.html